Renames the product to "Quick Unzip" (CLI command stays `qunzip`),
splits the Windows build into separate console (qunzip.exe) and GUI
(QuickUnzip.exe) binaries to eliminate the double-click console flash,
adds an Add-to-PATH toggle in settings, ships the rebuilt Inno Setup
installer for both binaries, fixes the auto-close race for small
archives, and cleans up compile warnings + adds detekt.

+++ src/mingwX64Main/kotlin/qunzip/platform/WindowsFileSystemRepository.kt
@@ -37,7 +37,7 @@ class WindowsFileSystemRepository(
         FileInfo(
             path = path,
             size = statBuf.st_size.toLong(),
-            lastModified = kotlinx.datetime.Instant.fromEpochSeconds(statBuf.st_mtime.toLong()),
+            lastModified = kotlinx.datetime.Instant.fromEpochSeconds(statBuf.st_mtime),
             isReadable = isReadable(path),
             isDirectory = (statBuf.st_mode.toInt() and S_IFDIR) != 0
         )
